Rename events to result since that's all we have now
