+++ Lib/fontbakery/reporters/serialize.py
@@ -10,7 +10,7 @@ domains as well.
 Domain specific knowledge should be encoded only in the Profile (Checks,
 Conditions) and MAYBE in *customized* reporters e.g. subclasses.
 """
-from fontbakery.events import CheckResult
+from fontbakery.result import CheckResult
 from fontbakery.reporters import FontbakeryReporter
