+++ Lib/fontbakery/result.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 from typing import Optional
 
 from fontbakery.section import Section
-from fontbakery.status import Status, ERROR, PASS
+from fontbakery.status import Status, ERROR
 from fontbakery.callable import FontBakeryCheck
 from fontbakery.message import Message
